opti code and start reconection
