+++ frontend/src/client/chatClient.ts
@@ -2,7 +2,7 @@ import { getSocket } from './socketClient.js';
 import { getUserIdFromToken } from './loginClient.js'
 
 export function init() {
-    const token = localStorage.getItem('token'); // TOKEN LINK FROM THE USER CONNECTED
+    const token = localStorage.getItem('token');
     const chat = document.getElementById('chat-container') as HTMLDivElement | null;
     
     if (!chat) 
@@ -17,9 +17,11 @@ export function init() {
         try {
             var socket = getSocket(0);
             if (socket)
-                socket.emit('message', msg, getUserIdFromToken(), 'toma');
+                socket.emit('message', msg, getUserIdFromToken(), 'toma'/* change by targe name */);
             else
                 console.error(`error 404 : socket not found !`);
+            
+
             // console.log('Start sended message !');
             // const res = await fetch('http://localhost:3001/api/user/priv-msg/toma', {
             // method: 'POST',
